[reports] int -> long для объектных полей в отчетах

+++ server/src/main/java/lsfusion/server/data/type/ObjectType.java
@@ -95,7 +95,7 @@ public class ObjectType extends AbstractType<Long> {
     public int getMinimumWidth() { return getPreferredWidth(); }
 
     public boolean fillReportDrawField(ReportDrawField reportField) {
-        reportField.valueClass = Integer.class;
+        reportField.valueClass = Long.class;
         reportField.alignment = HorizontalAlignEnum.RIGHT.getValue();
         return true;
     }
